# [window] Optimization aggregation window by segment tree