… velox/exec/CMakeLists.txt
… velox/exec/SegmentTreeAggregate.cpp
… velox/exec/SegmentTreeAggregate.h
… velox/exec/Window.h
… velox/exec/WindowFunction.h
… velox/functions/lib/window/tests/WindowTestBase.cpp
… velox/functions/lib/window/tests/WindowTestBase.h
… velox/functions/prestosql/window/CMakeLists.txt
… velox/functions/prestosql/window/benchmark/CMakeLists.txt
… velox/functions/prestosql/window/benchmark/WindowBenchmark.cpp
… velox/functions/prestosql/window/tests/AggregateWindowTest.cpp
… velox/vector/SelectivityVector.h