This optimization is based on https://www.vldb.org/pvldb/vol8/p1058-leis.pdf paper and is similar to https://duckdb.org/2021/10/13/windowing.html

1. Description
* When the window frame size is relatively large, such as window frame size >= 16, simple aggregation (each record aggregates its window data) will have a large number of repeated calculations.
* When the window is not incremental, for example the window frame start and frame end can be changed, or the aggregation function is difficult to remove records, incrementalAggregation cannot be used for optimization. Only simple aggregations can be used.

2. The optimization idea based on segment trees to reduce the repeated calculation of aggregate functions and reduce the time complexity of calculating a record from O(n) to O(log n) when the window frame size changes arbitrarily.

3. Optimize scenarios
* There are window aggregate functions.
* Window frame size >= 16 (meaning that the row size in one partition should to be at least 16), and as the window frame size becomes larger, the optimization becomes better.

4. The following are the benchmark results. The results show that as the window window frame size becomes larger, the optimization becomes more obvious. When the window frame size is 2048, the optimization is more than 10 times.

```
============================================================================
[...]/window/benchmark/WindowBenchmark.cpp     relative  time/iter   iters/s
============================================================================
rawRowFrame(count_frame_16_k_array)                       330.49ms      3.03
segmentOptRowFrame(count_frame_16_k_array)                313.07ms      3.19
rawRowFrame(count_frame_64_k_array)                       369.08ms      2.71
segmentOptRowFrame(count_frame_64_k_array)                292.25ms      3.42
rawRowFrame(count_frame_256_k_array)                      381.35ms      2.62
segmentOptRowFrame(count_frame_256_k_array)               301.02ms      3.32
rawRowFrame(count_frame_512_k_array)                      449.17ms      2.23
segmentOptRowFrame(count_frame_512_k_array)               353.11ms      2.83
rawRowFrame(count_frame_2048_k_array)                        1.45s   688.97m
segmentOptRowFrame(count_frame_2048_k_array)              298.44ms      3.35
rawRowFrame(sum_frame_16_k_array)                         309.85ms      3.23
segmentOptRowFrame(sum_frame_16_k_array)                  311.58ms      3.21
rawRowFrame(sum_frame_64_k_array)                         361.58ms      2.77
segmentOptRowFrame(sum_frame_64_k_array)                  316.90ms      3.16
rawRowFrame(sum_frame_256_k_array)                        602.75ms      1.66
segmentOptRowFrame(sum_frame_256_k_array)                 410.00ms      2.44
rawRowFrame(sum_frame_512_k_array)                           1.08s   926.15m
segmentOptRowFrame(sum_frame_512_k_array)                 429.23ms      2.33
rawRowFrame(sum_frame_2048_k_array)                          5.49s   181.99m
segmentOptRowFrame(sum_frame_2048_k_array)                414.17ms      2.41
rawRowFrame(avg_frame_16_k_array)                         398.75ms      2.51
segmentOptRowFrame(avg_frame_16_k_array)                  355.38ms      2.81
rawRowFrame(avg_frame_64_k_array)                         365.24ms      2.74
segmentOptRowFrame(avg_frame_64_k_array)                  350.43ms      2.85
rawRowFrame(avg_frame_256_k_array)                        580.88ms      1.72
segmentOptRowFrame(avg_frame_256_k_array)                 337.23ms      2.97
rawRowFrame(avg_frame_512_k_array)                        938.85ms      1.07
segmentOptRowFrame(avg_frame_512_k_array)                 373.71ms      2.68
rawRowFrame(avg_frame_2048_k_array)                          4.25s   235.56m
segmentOptRowFrame(avg_frame_2048_k_array)                381.15ms      2.62
rawRowFrame(min_frame_16_k_array)                         320.73ms      3.12
segmentOptRowFrame(min_frame_16_k_array)                  312.86ms      3.20
rawRowFrame(min_frame_64_k_array)                         358.26ms      2.79
segmentOptRowFrame(min_frame_64_k_array)                  308.58ms      3.24
rawRowFrame(min_frame_256_k_array)                        467.94ms      2.14
segmentOptRowFrame(min_frame_256_k_array)                 309.83ms      3.23
rawRowFrame(min_frame_512_k_array)                        665.46ms      1.50
segmentOptRowFrame(min_frame_512_k_array)                 322.60ms      3.10
rawRowFrame(min_frame_2048_k_array)                          2.31s   432.04m
segmentOptRowFrame(min_frame_2048_k_array)                358.51ms      2.79
rawRowFrame(max_frame_16_k_array)                         327.33ms      3.05
segmentOptRowFrame(max_frame_16_k_array)                  334.73ms      2.99
rawRowFrame(max_frame_64_k_array)                         342.32ms      2.92
segmentOptRowFrame(max_frame_64_k_array)                  402.26ms      2.49
rawRowFrame(max_frame_256_k_array)                        552.13ms      1.81
segmentOptRowFrame(max_frame_256_k_array)                 298.60ms      3.35
rawRowFrame(max_frame_512_k_array)                        590.44ms      1.69
segmentOptRowFrame(max_frame_512_k_array)                 302.45ms      3.31
rawRowFrame(max_frame_2048_k_array)                          2.20s   455.02m
segmentOptRowFrame(max_frame_2048_k_array)                385.36ms      2.59
rawRowFrame(stddev_frame_16_k_array)                      319.41ms      3.13
segmentOptRowFrame(stddev_frame_16_k_array)               434.15ms      2.30
rawRowFrame(stddev_frame_64_k_array)                      378.43ms      2.64
segmentOptRowFrame(stddev_frame_64_k_array)               552.84ms      1.81
rawRowFrame(stddev_frame_256_k_array)                        1.08s   927.33m
segmentOptRowFrame(stddev_frame_256_k_array)              679.89ms      1.47
rawRowFrame(stddev_frame_512_k_array)                        1.07s   931.61m
segmentOptRowFrame(stddev_frame_512_k_array)              787.58ms      1.27
rawRowFrame(stddev_frame_2048_k_array)                       3.77s   265.07m
segmentOptRowFrame(stddev_frame_2048_k_array)             849.46ms      1.18
```