@shenh062326 2024-03-13
> @shenh062326 Curious if this is based on https://www.vldb.org/pvldb/vol8/p1058-leis.pdf paper and similar to https://duckdb.org/2021/10/13/windowing.html

Yes, the optimization idea is base on it.

> CI is red. Would you take a look?

OK.