@aditi-pandit 2024-03-21
@shenh062326 : Thanks for adding this optimization. Had a few request for you to help review the code.

- Please can you write a small design for SegmentTree describing the structure and the functions you have written.
- Can SegmentTree logic be implemented in a common library in Velox code ? That could help in other places as well. If that is the case then it would be great to write a class for it with its tests so that we can validate its correctness.
- The code here seems to have this config flag-driven. Generally Velox follows the philosophy of runtime optimizations so make a decision as you start evaluating. Could we trigger the SegmentTree based evaluation at runtime based on some criteria ?