@aditi-pandit 2024-04-02
> @aditi-pandit Can you take a look at this PR.

@shenh062326 : Would be great if you could add some unit tests for the segment tree structure by itself as well. That would help me validate your code overall.

I'll spend more time on this in the next few days.