@aditi-pandit 2024-04-02
Thanks @shenh062326 for working on this.

> > * Please can you write a small design doc for SegmentTree describing the structure and the functions you have written.
> 
> Add a design doc for SegmentTree: https://docs.google.com/document/d/126MIwwKJA_557_UUDy3QRqbO79ta0veTLZgKGObwRFo/edit

Please can you give me permissions to comment on this doc.

> 
> > * Can SegmentTree logic be implemented in a common library in Velox code ? That could help in other places as well. If that is the case then it would be great to write a class for it with its tests so that we can validate its correctness.
> 
> I'm not sure if segment tree will be used elsewhere, maybe it can be implemented in a common library in Velox in the future when this situation arises.

I still feel it will be better to separate SegmentTree into its own class with an API, so that we can test it independently from WindowFunctions. WindowFunctions might not be able to test all its corner cases in a way that ensures we got coverage.
wdyt ?
> 
> > * The code here seems to have this config flag-driven. Generally Velox follows the philosophy of runtime optimizations so make a decision as you start evaluating. Could we trigger the SegmentTree based evaluation at runtime based on some criteria ?
> 
> Currently, for each window partition, choose segmentTree algorithm rather than simpleAggregation if:
> 
> * If the window frame start and frame end is constant value,  and frame size is larger than min_frame_size_use_segment_tree (default value 16),

> * If the window frame is not constant, compute the average frame size for the input rows, if the average frame size is larger than min_frame_size_use_segment_tree (default value 16),
> 

This sounds good.

> If min_frame_size_use_segment_tree no need, we can use a constant value instead.
Did you get this number from the benchmark runs ? In any case, having a config for it is fine imo.