@shenh062326 2024-04-07
> Thanks @shenh062326 for working on this.
> 
> Please can you give me permissions to comment on this doc.

Done.

> 
> I still feel it will be better to separate SegmentTree into its own class with an API, so that we can test it independently from WindowFunctions. WindowFunctions might not be able to test all its corner cases in a way that ensures we got coverage. wdyt ?
> 

OK, I will separate SegmentTree.

> Did you get this number from the benchmark runs ? In any case, having a config for it is fine imo.

Yes, I get the number from  benchmark. I will keep one config.