@shenh062326 2024-04-10
1. Rebase master to fix conflict.
2. Separate SegmentTree into its own class.
3. Remove config min_frame_size_use_segment_tree and add kMinFrameSizeUseSegmentTree.