@aditi-pandit 2024-04-18
Thanks @shenh062326. I left some questions on your doc.

Your code looks good overall. Though it would be great to separate this code into 2 PRs for easier review : 

- Segment tree class and tests in the first PR.
- The rest of the changes in this PR in a 2nd one.